relocate check command error handling to commands package
